pyproject.toml
[project]
name = "hackithon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.115.12",
    "fastapi-mqtt>=2.2.0",
    "openai>=1.82.0",
    "openapi>=2.0.0",
    "pathlib>=1.0.1",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
